+++ Source/WebCore/rendering_scion/RenderBlockFlow.swift
@@ -3411,7 +3411,7 @@ class RenderBlockFlowWrapper: RenderBlockWrapper {
           value: (fontMetrics.intAscent()
             + (lineHeight(
               firstLine: true, direction: lineDirection,
-              linePositionMode: .PositionOfInteriorLineBoxes) - fontMetrics.intHeight()) / 2
+              linePositionMode: .PositionOfInteriorLineBoxes) - Int32(fontMetrics.intHeight())) / 2
             + (lineDirection == .HorizontalLine
               ? borderTop() + paddingTop() : borderRight() + paddingRight()))
             .toInt())
@@ -5434,8 +5434,8 @@ class RenderBlockFlowWrapper: RenderBlockWrapper {
 
     // Make an adjustment to align with the cap height of a theoretical block line.
     let adjustment =
-      fontMetrics.intAscent() + (heightOfLine - fontMetrics.intHeight()) / 2
-      - fontMetrics.intCapHeight() - beforeMarginBorderPadding
+      fontMetrics.intAscent() + (heightOfLine - Int32(fontMetrics.intHeight())) / 2
+      - Int32(fontMetrics.intCapHeight()) - beforeMarginBorderPadding
     logicalTopOffset += adjustment
 
     // For sunken and raised caps, we have to make some adjustments. Test if we're sunken or raised (dropHeightDelta will be
